[REFACTORING] - Skiller ut profilering
Profilering er i første omgang kun snakk om 2 klasser, men her ligger det et større potensialet.
Det er mer logikk som bør hentes ut fra ulike services/utils, og på sikt bør en forsøke å dele opp transaksjonsgrensene også.
