+++ src/main/java/no/nav/fo/veilarbregistrering/utils/FunksjonelleMetrikker.java
@@ -3,6 +3,7 @@ package no.nav.fo.veilarbregistrering.utils;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import no.nav.fo.veilarbregistrering.domain.*;
+import no.nav.fo.veilarbregistrering.profilering.Profilering;
 import no.nav.metrics.Event;
 import no.nav.metrics.MetricsFactory;
